[Ref] Move vocabulary importing-related components into separate directory.
