+++ src/component/vocabulary/importing/ImportVocabularyRoute.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import IfUserIsEditor from "../authorization/IfUserIsEditor";
+import IfUserIsEditor from "../../authorization/IfUserIsEditor";
 import ImportVocabularyPage from "./ImportVocabularyPage";
 
 const ImportVocabularyRoute: React.FC = () => (
